requirements.txt
langchain==0.3.0
langchain-anthropic==0.3.0
langchain-mongodb==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
langgraph==0.2.28
pymongo[srv]==4.7.0
streamlit==1.38.0
pypdf==4.3.0
python-dotenv==1.0.0
sentence-transformers==3.0.0
tiktoken==0.7.0
